Start refactor of RYD
